Address topology interaction review feedback
