+++ ChartForgeX/ChartForgeX.csproj
@@ -40,5 +40,6 @@
   <ItemGroup>
     <None Include="..\README.nuget.md" Pack="true" PackagePath="README.md" />
     <None Include="..\assets\ChartForgeX.png" Pack="true" PackagePath="\" />
+    <EmbeddedResource Include="Topology\Assets\topology-interaction.js" LogicalName="ChartForgeX.Topology.Assets.topology-interaction.js" />
   </ItemGroup>
 </Project>
