+++ ChartForgeX/Topology/Assets/topology-interaction.js
@@ -1,8 +1,588 @@
-namespace ChartForgeX.Topology;
+(() => {
+  for (const wrapper of document.querySelectorAll('.cfx-topology-wrapper[data-cfx-interactive="true"]')) {
+    if (wrapper.getAttribute('data-cfx-runtime-bound') === 'true') continue;
+    wrapper.setAttribute('data-cfx-runtime-bound', 'true');
+    const selectables = '[data-cfx-role="topology-node"],[data-cfx-role="topology-edge"],[data-cfx-role="topology-group"]';
+    const viewportControls = wrapper.getAttribute('data-cfx-viewport-controls') === 'true';
+    const exportControls = wrapper.getAttribute('data-cfx-export-controls') === 'true';
+    const fullscreenControl = wrapper.getAttribute('data-cfx-fullscreen-control') === 'true';
+    const scenarioControls = wrapper.getAttribute('data-cfx-scenario-controls') === 'true';
+    const scenarioControlMode = wrapper.getAttribute('data-cfx-scenario-control-mode') || 'buttons';
+    const forceGraphControls = wrapper.getAttribute('data-cfx-force-graph-controls') === 'true';
+    const scenarioUrlState = wrapper.getAttribute('data-cfx-scenario-url-state') === 'true';
+    const scenarioPanel = wrapper.querySelector('[data-cfx-topology-scenario-panel]');
+    const selectionPanel = wrapper.querySelector('[data-cfx-topology-selection-panel]');
+    const forceGraphPanel = wrapper.querySelector('[data-cfx-force-graph-panel]');
+    const syncEnabled = wrapper.getAttribute('data-cfx-sync-enabled') === 'true';
+    const syncGroup = wrapper.getAttribute('data-cfx-sync-group') || '';
+    const viewport = wrapper.querySelector('.cfx-topology-viewport') || wrapper;
+    const topologyRoot = wrapper.querySelector('[data-cfx-role="topology"]');
+    const topologySvg = () => {
+      const root = wrapper.querySelector('[data-cfx-role="topology"]');
+      return root ? root.closest('svg') : (viewport.querySelector(':scope > svg') || wrapper.querySelector('svg'));
+    };
+    let applyingSync = false;
+    let scenarioPlayback = null;
+    let forceGraphMotionTimer = null;
+    const scenarioUrlKey = name => {
+      const chartId = attr(wrapper, 'data-chart-id').replace(/[^a-z0-9_.-]+/gi, '-').replace(/^-+|-+$/g, '') || 'topology';
+      return 'cfx-' + chartId + '-' + name;
+    };
+    const scenarioUrlParam = name => scenarioUrlState ? (() => { try { const params = new URL(window.location.href).searchParams; return params.get(scenarioUrlKey(name)) || params.get(name) || ''; } catch { return ''; } })() : '';
+    const syncScenarioUrl = (scenarioId, stepIndex) => {
+      if (!scenarioUrlState || !window.history || !window.history.replaceState) return;
+      try { const url = new URL(window.location.href); const scenarioKey = scenarioUrlKey('scenario'); const stepKey = scenarioUrlKey('scenarioStep'); scenarioId ? url.searchParams.set(scenarioKey, scenarioId) : url.searchParams.delete(scenarioKey); stepIndex === undefined || stepIndex === null || stepIndex === '' ? url.searchParams.delete(stepKey) : url.searchParams.set(stepKey, stepIndex); url.searchParams.delete('scenario'); url.searchParams.delete('scenarioStep'); window.history.replaceState(window.history.state, '', url); } catch { }
+    };
+    const attr = (element, name) => element.getAttribute(name) || '';
+    const toCamel = value => value.replace(/-([a-z0-9])/g, (_, ch) => ch.toUpperCase());
+    const collectPrefixed = (element, prefix) => {
+      const values = {};
+      for (const attribute of element.attributes) {
+        if (attribute.name.startsWith(prefix)) values[toCamel(attribute.name.slice(prefix.length))] = attribute.value;
+      }
+      return values;
+    };
+    const unique = values => Array.from(new Set(values.filter(value => value)));
+    const scenarioIdTokens = value => unique(String(value || '').split(/[,\s]+/).map(item => item.trim()));
+    const scenarioIds = element => (attr(element, 'data-scenario-ids') || '').split(/\s+/).filter(value => value);
+    const scenarioStepIndices = (element, scenarioId) => (attr(element, 'data-scenario-step-indices') || '').split(/\s+/).filter(value => value.startsWith(scenarioId + ':')).flatMap(value => value.slice(scenarioId.length + 1).split(',').filter(index => index));
+    const findScenarioButton = scenarioId => Array.from(wrapper.querySelectorAll('[data-cfx-topology-scenario],[data-cfx-topology-scenario-toggle]')).find(button => (attr(button, 'data-cfx-topology-scenario') || attr(button, 'data-cfx-topology-scenario-toggle')) === scenarioId) || null;
+    const scenarioSummaries = () => {
+      try {
+        const summaries = JSON.parse(attr(topologyRoot || wrapper, 'data-cfx-scenarios') || '[]');
+        return Array.isArray(summaries) ? summaries : [];
+      } catch {
+        return [];
+      }
+    };
+    const findScenarioSummary = scenarioId => scenarioSummaries().find(scenario => scenario && scenario.id === scenarioId) || null;
+    const scenarioSteps = button => {
+      try {
+        const steps = JSON.parse(attr(button, 'data-cfx-scenario-steps') || '[]');
+        return Array.isArray(steps) ? steps : [];
+      } catch {
+        return [];
+      }
+    };
+    const scenarioMetadata = button => {
+      try {
+        const metadata = JSON.parse(attr(button, 'data-cfx-scenario-metadata') || '{}');
+        return metadata && typeof metadata === 'object' && !Array.isArray(metadata) ? metadata : {};
+      } catch {
+        return {};
+      }
+    };
+    const scenarioRoute = scenarioId => {
+      if (!scenarioId) return null;
+      const scenarioButton = findScenarioButton(scenarioId);
+      const scenarioSummary = scenarioButton ? null : findScenarioSummary(scenarioId);
+      if (!scenarioButton && !scenarioSummary) return null;
+      const steps = scenarioButton ? scenarioSteps(scenarioButton) : (Array.isArray(scenarioSummary.steps) ? scenarioSummary.steps : []);
+      const metadata = scenarioButton ? scenarioMetadata(scenarioButton) : (scenarioSummary.metadata && typeof scenarioSummary.metadata === 'object' && !Array.isArray(scenarioSummary.metadata) ? scenarioSummary.metadata : {});
+      const nodeIds = new Set();
+      const edgeIds = new Set();
+      const groupIds = new Set();
+      wrapper.querySelectorAll('[data-cfx-role="topology-node"][data-scenario-ids]').forEach(node => {
+        if (!scenarioIds(node).includes(scenarioId)) return;
+        nodeIds.add(attr(node, 'data-node-id'));
+        groupIds.add(attr(node, 'data-group-id'));
+      });
+      wrapper.querySelectorAll('[data-cfx-role="topology-edge"][data-scenario-ids]').forEach(edge => {
+        if (!scenarioIds(edge).includes(scenarioId)) return;
+        edgeIds.add(attr(edge, 'data-edge-id'));
+        nodeIds.add(attr(edge, 'data-source-node-id'));
+        nodeIds.add(attr(edge, 'data-target-node-id'));
+        groupIds.add(attr(edge, 'data-source-group-id'));
+        groupIds.add(attr(edge, 'data-target-group-id'));
+      });
+      return {
+        scenarioId,
+        button: scenarioButton,
+        label: scenarioButton ? attr(scenarioButton, 'data-cfx-scenario-label') || scenarioButton.textContent.trim() : scenarioSummary.label || scenarioId,
+        description: scenarioButton ? attr(scenarioButton, 'data-cfx-scenario-description') : scenarioSummary.description || '',
+        color: scenarioButton ? attr(scenarioButton, 'data-cfx-scenario-color') : scenarioSummary.color || '',
+        stepCount: scenarioButton ? Number(attr(scenarioButton, 'data-cfx-scenario-step-count') || '0') : Number(scenarioSummary.stepCount || steps.length || 0),
+        metadata,
+        steps,
+        nodeIds,
+        edgeIds,
+        groupIds
+      };
+    };
+    const renderScenarioPanel = route => {
+      if (!scenarioPanel) return;
+      const title = scenarioPanel.querySelector('[data-cfx-scenario-panel-title]');
+      const meta = scenarioPanel.querySelector('[data-cfx-scenario-panel-meta]');
+      const stepsList = scenarioPanel.querySelector('[data-cfx-scenario-panel-steps]');
+      if (!route) {
+        scenarioPanel.removeAttribute('data-cfx-panel-active-scenario');
+        scenarioPanel.style.removeProperty('--cfx-topology-scenario-color');
+        if (title) title.textContent = 'All';
+        if (meta) meta.textContent = 'All topology routes visible';
+        if (stepsList) stepsList.replaceChildren();
+        return;
+      }
 
-public sealed partial class TopologyHtmlRenderer
-{
-    private static string InteractionScriptPart2() => """
+      scenarioPanel.setAttribute('data-cfx-panel-active-scenario', route.scenarioId);
+      if (route.color) scenarioPanel.style.setProperty('--cfx-topology-scenario-color', route.color);
+      else scenarioPanel.style.removeProperty('--cfx-topology-scenario-color');
+      if (title) title.textContent = route.label || route.scenarioId;
+      if (meta) {
+        const metadata = Object.entries(route.metadata).slice(0, 2).map(([key, value]) => key + ': ' + value);
+        const summary = [route.description, route.stepCount ? route.stepCount + ' steps' : '', ...metadata].filter(value => value);
+        meta.textContent = summary.join(' / ');
+      }
+      if (stepsList) {
+        stepsList.replaceChildren();
+        route.steps.forEach(step => {
+          const item = document.createElement('li');
+          item.setAttribute('data-cfx-scenario-step-kind', step.kind || '');
+          item.setAttribute('data-cfx-scenario-step-id', step.id || '');
+          item.setAttribute('data-cfx-scenario-step-index', String((Number(step.index) || 0) + 1));
+          item.setAttribute('role', 'button');
+          item.setAttribute('tabindex', '0');
+          item.textContent = step.label || step.id || '';
+          stepsList.appendChild(item);
+        });
+      }
+    };
+    const clearScenarioPreview = () => {
+      wrapper.style.removeProperty('--cfx-topology-preview-scenario-color');
+      wrapper.querySelectorAll('.cfx-topology-html-scenario-preview,.cfx-topology-html-scenario-preview-muted').forEach(item => {
+        item.classList.remove('cfx-topology-html-scenario-preview', 'cfx-topology-html-scenario-preview-muted');
+      });
+    };
+    const previewScenario = scenarioId => {
+      clearScenarioPreview();
+      if (!scenarioId) return;
+      const route = scenarioRoute(scenarioId);
+      if (!route) return;
+      if (route.color) wrapper.style.setProperty('--cfx-topology-preview-scenario-color', route.color);
+      wrapper.querySelectorAll('[data-cfx-role="topology-edge"]').forEach(edge => {
+        const active = route.edgeIds.has(attr(edge, 'data-edge-id'));
+        edge.classList.toggle('cfx-topology-html-scenario-preview', active);
+        edge.classList.toggle('cfx-topology-html-scenario-preview-muted', !active);
+      });
+      wrapper.querySelectorAll('[data-cfx-role="topology-edge-label"]').forEach(label => {
+        const active = route.edgeIds.has(attr(label, 'data-edge-id'));
+        label.classList.toggle('cfx-topology-html-scenario-preview', active);
+        label.classList.toggle('cfx-topology-html-scenario-preview-muted', !active);
+      });
+      wrapper.querySelectorAll('[data-cfx-role="topology-node"]').forEach(node => {
+        const active = route.nodeIds.has(attr(node, 'data-node-id'));
+        node.classList.toggle('cfx-topology-html-scenario-preview', active);
+        node.classList.toggle('cfx-topology-html-scenario-preview-muted', !active);
+      });
+      wrapper.querySelectorAll('[data-cfx-role="topology-node-status"]').forEach(status => {
+        const active = route.nodeIds.has(attr(status, 'data-node-id'));
+        status.classList.toggle('cfx-topology-html-scenario-preview', active);
+        status.classList.toggle('cfx-topology-html-scenario-preview-muted', !active);
+      });
+      wrapper.querySelectorAll('[data-cfx-role="topology-group"]').forEach(group => {
+        const active = route.groupIds.has(attr(group, 'data-group-id'));
+        group.classList.toggle('cfx-topology-html-scenario-preview', active);
+        group.classList.toggle('cfx-topology-html-scenario-preview-muted', !active);
+      });
+      wrapper.dispatchEvent(new CustomEvent('cfx-topology-scenario-preview', { bubbles: true, detail: { chartId: attr(wrapper, 'data-chart-id'), scenarioId, label: route.label, description: route.description, color: route.color, stepCount: route.stepCount, metadata: route.metadata, steps: route.steps, nodeIds: Array.from(route.nodeIds), edgeIds: Array.from(route.edgeIds), groupIds: Array.from(route.groupIds) } }));
+    };
+    const updateScenarioStepControls = playing => {
+      const hasRoute = !!scenarioRoute(attr(wrapper, 'data-cfx-active-scenario'));
+      wrapper.querySelectorAll('[data-cfx-scenario-step-control]').forEach(button => {
+        button.disabled = !hasRoute;
+        if (attr(button, 'data-cfx-scenario-step-control') === 'play') button.setAttribute('aria-pressed', playing ? 'true' : 'false');
+      });
+    };
+    const clearScenarioStep = () => {
+      wrapper.removeAttribute('data-cfx-active-scenario-step');
+      wrapper.querySelectorAll('.cfx-topology-html-scenario-step-active,.cfx-topology-html-scenario-step-muted').forEach(item => item.classList.remove('cfx-topology-html-scenario-step-active', 'cfx-topology-html-scenario-step-muted'));
+      if (scenarioPanel) scenarioPanel.querySelectorAll('[data-cfx-scenario-step-id]').forEach(item => item.removeAttribute('aria-current'));
+      updateScenarioStepControls(false);
+    };
+    const stopScenarioPlayback = () => {
+      if (scenarioPlayback) window.clearInterval(scenarioPlayback);
+      scenarioPlayback = null;
+      updateScenarioStepControls(false);
+    };
+    const setScenarioStep = (stepIndex, keepPlaying, emit = true, sync = true) => {
+      const route = scenarioRoute(attr(wrapper, 'data-cfx-active-scenario'));
+      if (!route || !route.steps.length) return;
+      const requested = Number(stepIndex);
+      const index = clamp(Number.isFinite(requested) ? requested : 0, 0, route.steps.length - 1);
+      wrapper.setAttribute('data-cfx-active-scenario-step', String(index));
+      wrapper.querySelectorAll('[data-cfx-role="topology-edge"],[data-cfx-role="topology-node"]').forEach(item => {
+        const routeMember = item.classList.contains('cfx-topology-html-scenario-active');
+        const current = routeMember && scenarioStepIndices(item, route.scenarioId).includes(String(index));
+        item.classList.toggle('cfx-topology-html-scenario-step-active', current);
+        item.classList.toggle('cfx-topology-html-scenario-step-muted', routeMember && !current);
+      });
+      if (scenarioPanel) scenarioPanel.querySelectorAll('[data-cfx-scenario-step-id]').forEach(item => {
+        if (attr(item, 'data-cfx-scenario-step-index') === String(index + 1)) item.setAttribute('aria-current', 'step');
+        else item.removeAttribute('aria-current');
+      });
+      if (!keepPlaying) stopScenarioPlayback();
+      if (emit) wrapper.dispatchEvent(new CustomEvent('cfx-topology-scenario-step', { bubbles: true, detail: { chartId: attr(wrapper, 'data-chart-id'), scenarioId: route.scenarioId, index, step: route.steps[index] } }));
+      syncScenarioUrl(route.scenarioId, index);
+      if (sync) emitSync('scenario-step', { scenarioId: route.scenarioId, index });
+    };
+    const stepScenario = delta => {
+      const route = scenarioRoute(attr(wrapper, 'data-cfx-active-scenario'));
+      if (!route) return;
+      const current = Number(wrapper.getAttribute('data-cfx-active-scenario-step') || (delta > 0 ? '-1' : route.steps.length));
+      setScenarioStep(current + delta, false);
+    };
+    const playScenario = () => {
+      const route = scenarioRoute(attr(wrapper, 'data-cfx-active-scenario'));
+      if (!route || !route.steps.length) return;
+      if (scenarioPlayback) {
+        stopScenarioPlayback();
+        return;
+      }
+      updateScenarioStepControls(true);
+      let index = Number(wrapper.getAttribute('data-cfx-active-scenario-step') || '-1') + 1;
+      if (!Number.isFinite(index) || index >= route.steps.length) index = 0;
+      scenarioPlayback = window.setInterval(() => {
+        if (index >= route.steps.length) {
+          stopScenarioPlayback();
+          return;
+        }
+        setScenarioStep(index++, true);
+      }, 900);
+      setScenarioStep(index++, true);
+    };
+    const copyScenarioLink = () => { syncScenarioUrl(attr(wrapper, 'data-cfx-active-scenario') || attr(wrapper, 'data-cfx-active-scenarios'), wrapper.getAttribute('data-cfx-active-scenario-step') || ''); const url = window.location.href; if (navigator.clipboard && navigator.clipboard.writeText) navigator.clipboard.writeText(url).catch(() => {}); wrapper.dispatchEvent(new CustomEvent('cfx-topology-scenario-link', { bubbles: true, detail: { chartId: attr(wrapper, 'data-chart-id'), url } })); };
+    const emitFullscreenState = () => wrapper.dispatchEvent(new CustomEvent('cfx-topology-fullscreen', { bubbles: true, detail: { chartId: attr(wrapper, 'data-chart-id'), fullscreen: document.fullscreenElement === wrapper } }));
+    const setFullscreenState = () => { wrapper.setAttribute('data-cfx-topology-fullscreen', document.fullscreenElement === wrapper ? 'true' : 'false'); emitFullscreenState(); };
+    const toggleFullscreen = () => {
+      if (!document.fullscreenElement && wrapper.requestFullscreen) wrapper.requestFullscreen().catch(() => {});
+      else if (document.fullscreenElement === wrapper && document.exitFullscreen) document.exitFullscreen().catch(() => {});
+    };
+    const clamp = (value, min, max) => Math.max(min, Math.min(max, value));
+    const numberAttr = (name, fallback) => {
+      const value = Number(wrapper.getAttribute(name) || '');
+      return Number.isFinite(value) ? value : fallback;
+    };
+    const viewportState = () => ({
+      zoom: numberAttr('data-cfx-topology-zoom', 1),
+      panX: numberAttr('data-cfx-topology-pan-x', 0),
+      panY: numberAttr('data-cfx-topology-pan-y', 0)
+    });
+    const emitSync = (action, payload) => {
+      if (!syncEnabled || !syncGroup || applyingSync) return;
+      const detail = Object.assign({ chartId: attr(wrapper, 'data-chart-id'), group: syncGroup, action }, payload || {});
+      wrapper.dispatchEvent(new CustomEvent('cfx-topology-sync', { bubbles: true, detail }));
+      document.querySelectorAll('.cfx-topology-wrapper[data-cfx-interactive="true"][data-cfx-sync-enabled="true"]').forEach(peer => {
+        if (peer === wrapper || peer.getAttribute('data-cfx-sync-group') !== syncGroup) return;
+        peer.dispatchEvent(new CustomEvent('cfx-topology-apply-sync', { detail }));
+      });
+    };
+    const applyViewport = state => {
+      if (!viewportControls) return;
+      const next = {
+        zoom: clamp(Number(state.zoom) || 1, 0.5, 4),
+        panX: clamp(Number(state.panX) || 0, -4000, 4000),
+        panY: clamp(Number(state.panY) || 0, -4000, 4000)
+      };
+      wrapper.setAttribute('data-cfx-topology-zoom', next.zoom.toFixed(3));
+      wrapper.setAttribute('data-cfx-topology-pan-x', next.panX.toFixed(1));
+      wrapper.setAttribute('data-cfx-topology-pan-y', next.panY.toFixed(1));
+      wrapper.style.setProperty('--cfx-topology-zoom', next.zoom.toFixed(3));
+      wrapper.style.setProperty('--cfx-topology-pan-x', next.panX.toFixed(1) + 'px');
+      wrapper.style.setProperty('--cfx-topology-pan-y', next.panY.toFixed(1) + 'px');
+    };
+    const viewportMetrics = () => {
+      const svg = topologySvg();
+      const viewBox = svg && svg.viewBox && svg.viewBox.baseVal ? svg.viewBox.baseVal : null;
+      const rect = viewport.getBoundingClientRect ? viewport.getBoundingClientRect() : { width: 0, height: 0 };
+      const styles = window.getComputedStyle ? window.getComputedStyle(viewport) : null;
+      const paddingTop = styles ? Number.parseFloat(styles.paddingTop || '0') || 0 : 0;
+      const paddingLeft = styles ? Number.parseFloat(styles.paddingLeft || '0') || 0 : 0;
+      const paddingRight = styles ? Number.parseFloat(styles.paddingRight || '0') || 0 : 0;
+      const paddingBottom = styles ? Number.parseFloat(styles.paddingBottom || '0') || 0 : 0;
+      const width = Math.max(1, rect.width - paddingLeft - paddingRight);
+      const availableHeight = Math.max(1, rect.height - paddingTop - paddingBottom);
+      const svgWidth = Math.max(1, width);
+      const svgHeight = viewBox && viewBox.width > 0 && viewBox.height > 0 ? svgWidth * (viewBox.height / viewBox.width) : Math.max(1, svg ? svg.getBoundingClientRect().height : availableHeight);
+      return { rect, paddingTop, paddingLeft, width, availableHeight, svgWidth, svgHeight };
+    };
+    const emitViewport = () => {
+      wrapper.dispatchEvent(new CustomEvent('cfx-topology-viewport', { bubbles: true, detail: { chartId: attr(wrapper, 'data-chart-id'), state: viewportState() } }));
+      emitSync('viewport', { state: viewportState() });
+    };
+    const zoomBy = (factor, origin) => {
+      const state = viewportState();
+      const nextZoom = clamp(state.zoom * factor, 0.5, 4);
+      const ratio = nextZoom / state.zoom;
+      const metrics = viewportMetrics();
+      const localX = origin && Number.isFinite(origin.clientX) ? origin.clientX - metrics.rect.left : metrics.rect.width / 2;
+      const localY = origin && Number.isFinite(origin.clientY) ? origin.clientY - metrics.rect.top : metrics.rect.height / 2;
+      const centerX = metrics.rect.width / 2;
+      const centerY = metrics.rect.height / 2;
+      const panX = localX - centerX - ratio * (localX - centerX - state.panX);
+      const panY = localY - centerY - ratio * (localY - centerY - state.panY);
+      markForceGraphMoving();
+      applyViewport({ zoom: nextZoom, panX, panY });
+      emitViewport();
+    };
+    const fitViewport = () => {
+      const metrics = viewportMetrics();
+      const scale = clamp(Math.min(1, metrics.width / metrics.svgWidth, metrics.availableHeight / metrics.svgHeight), 0.5, 1);
+      applyViewport({ zoom: scale, panX: 0, panY: 0 });
+      wrapper.dispatchEvent(new CustomEvent('cfx-topology-fit', { bubbles: true, detail: { chartId: attr(wrapper, 'data-chart-id'), state: viewportState() } }));
+      emitViewport();
+    };
+    const markForceGraphMoving = () => {
+      if (!forceGraphControls || wrapper.getAttribute('data-cfx-force-hide-moving-edges') !== 'true') return;
+      wrapper.setAttribute('data-cfx-force-moving-edges', 'true');
+      wrapper.querySelectorAll('[data-cfx-role="topology-edge"]').forEach(edge => edge.classList.add('cfx-topology-html-force-moving'));
+      if (forceGraphMotionTimer) window.clearTimeout(forceGraphMotionTimer);
+      forceGraphMotionTimer = window.setTimeout(() => {
+        wrapper.removeAttribute('data-cfx-force-moving-edges');
+        wrapper.querySelectorAll('.cfx-topology-html-force-moving').forEach(edge => edge.classList.remove('cfx-topology-html-force-moving'));
+      }, 280);
+    };
+    const resetViewport = () => {
+      applyViewport({ zoom: 1, panX: 0, panY: 0 });
+      wrapper.removeAttribute('data-cfx-topology-dragging');
+      wrapper.dispatchEvent(new CustomEvent('cfx-topology-reset', { bubbles: true, detail: { chartId: attr(wrapper, 'data-chart-id') } }));
+      emitViewport();
+    };
+    const setViewportMode = mode => {
+      const next = wrapper.getAttribute('data-cfx-topology-mode') === mode ? '' : mode;
+      if (next) wrapper.setAttribute('data-cfx-topology-mode', next);
+      else wrapper.removeAttribute('data-cfx-topology-mode');
+      wrapper.querySelectorAll('[data-cfx-topology-mode]').forEach(button => button.setAttribute('aria-pressed', attr(button, 'data-cfx-topology-mode') === next ? 'true' : 'false'));
+    };
+    const clearScenario = () => {
+      stopScenarioPlayback();
+      clearScenarioStep();
+      clearScenarioPreview();
+      wrapper.removeAttribute('data-cfx-active-scenario');
+      wrapper.removeAttribute('data-cfx-active-scenarios');
+      wrapper.style.removeProperty('--cfx-topology-active-scenario-color');
+      wrapper.querySelectorAll('.cfx-topology-html-scenario-active,.cfx-topology-html-scenario-muted').forEach(item => {
+        item.classList.remove('cfx-topology-html-scenario-active', 'cfx-topology-html-scenario-muted');
+        item.removeAttribute('data-cfx-active-scenario-step-indices');
+      });
+      wrapper.querySelectorAll('[data-cfx-topology-scenario]').forEach(button => button.setAttribute('aria-pressed', attr(button, 'data-cfx-topology-scenario') ? 'false' : 'true'));
+      wrapper.querySelectorAll('[data-cfx-topology-scenario-toggle]').forEach(input => input.checked = false);
+      updateScenarioStepControls(false);
+      renderScenarioPanel(null);
+      syncScenarioUrl('', '');
+    };
+    const setScenarioFilters = (scenarioIds, emit = true, sync = true) => {
+      clearScenario();
+      const ids = unique((scenarioIds || []).map(value => String(value || '').trim()));
+      if (!ids.length) {
+        if (emit) wrapper.dispatchEvent(new CustomEvent('cfx-topology-scenario-clear', { bubbles: true, detail: { chartId: attr(wrapper, 'data-chart-id') } }));
+        if (sync) emitSync('scenario-filters', { scenarioIds: [] });
+        return;
+      }
+
+      const routes = ids.map(scenarioRoute).filter(route => route);
+      if (!routes.length) {
+        if (emit) wrapper.dispatchEvent(new CustomEvent('cfx-topology-scenario-clear', { bubbles: true, detail: { chartId: attr(wrapper, 'data-chart-id') } }));
+        if (sync) emitSync('scenario-filters', { scenarioIds: [] });
+        return;
+      }
+
+      const nodeIds = new Set();
+      const edgeIds = new Set();
+      const groupIds = new Set();
+      routes.forEach(route => {
+        route.nodeIds.forEach(id => nodeIds.add(id));
+        route.edgeIds.forEach(id => edgeIds.add(id));
+        route.groupIds.forEach(id => groupIds.add(id));
+      });
+      wrapper.setAttribute('data-cfx-active-scenarios', routes.map(route => route.scenarioId).join(' '));
+      if (routes.length === 1) {
+        wrapper.setAttribute('data-cfx-active-scenario', routes[0].scenarioId);
+        if (routes[0].color) wrapper.style.setProperty('--cfx-topology-active-scenario-color', routes[0].color);
+        else wrapper.style.removeProperty('--cfx-topology-active-scenario-color');
+      } else {
+        wrapper.removeAttribute('data-cfx-active-scenario');
+        wrapper.style.removeProperty('--cfx-topology-active-scenario-color');
+      }
+      wrapper.querySelectorAll('[data-cfx-topology-scenario-toggle]').forEach(input => input.checked = routes.some(route => route.scenarioId === attr(input, 'data-cfx-topology-scenario-toggle')));
+      wrapper.querySelectorAll('[data-cfx-role="topology-edge"]').forEach(edge => {
+        const active = edgeIds.has(attr(edge, 'data-edge-id'));
+        edge.classList.toggle('cfx-topology-html-scenario-active', active);
+        edge.classList.toggle('cfx-topology-html-scenario-muted', !active);
+        if (active && routes.length === 1) edge.setAttribute('data-cfx-active-scenario-step-indices', scenarioStepIndices(edge, routes[0].scenarioId).join(' '));
+      });
+      wrapper.querySelectorAll('[data-cfx-role="topology-node"]').forEach(node => {
+        const active = nodeIds.has(attr(node, 'data-node-id'));
+        node.classList.toggle('cfx-topology-html-scenario-active', active);
+        node.classList.toggle('cfx-topology-html-scenario-muted', !active);
+        if (active && routes.length === 1) node.setAttribute('data-cfx-active-scenario-step-indices', scenarioStepIndices(node, routes[0].scenarioId).join(' '));
+      });
+      wrapper.querySelectorAll('[data-cfx-role="topology-edge-label"]').forEach(label => {
+        const active = edgeIds.has(attr(label, 'data-edge-id'));
+        label.classList.toggle('cfx-topology-html-scenario-active', active);
+        label.classList.toggle('cfx-topology-html-scenario-muted', !active);
+      });
+      wrapper.querySelectorAll('[data-cfx-role="topology-node-status"]').forEach(status => {
+        const active = nodeIds.has(attr(status, 'data-node-id'));
+        status.classList.toggle('cfx-topology-html-scenario-active', active);
+        status.classList.toggle('cfx-topology-html-scenario-muted', !active);
+      });
+      wrapper.querySelectorAll('[data-cfx-role="topology-group"]').forEach(group => {
+        const active = groupIds.has(attr(group, 'data-group-id'));
+        group.classList.toggle('cfx-topology-html-scenario-active', active);
+        group.classList.toggle('cfx-topology-html-scenario-muted', !active);
+      });
+      renderScenarioPanel(routes.length === 1 ? routes[0] : { ...routes[0], label: routes.length + ' routes enabled', description: routes.map(route => route.label).join(' / '), stepCount: routes.reduce((sum, route) => sum + route.stepCount, 0), steps: routes.flatMap(route => route.steps), metadata: {} });
+      updateScenarioStepControls(false);
+      syncScenarioUrl(routes.map(route => route.scenarioId).join(','), '');
+      if (emit) wrapper.dispatchEvent(new CustomEvent('cfx-topology-scenario-filter', { bubbles: true, detail: { chartId: attr(wrapper, 'data-chart-id'), scenarioIds: routes.map(route => route.scenarioId), routes: routes.map(route => ({ scenarioId: route.scenarioId, label: route.label, description: route.description, color: route.color, stepCount: route.stepCount })), nodeIds: Array.from(nodeIds), edgeIds: Array.from(edgeIds), groupIds: Array.from(groupIds) } }));
+      if (sync) emitSync('scenario-filters', { scenarioIds: routes.map(route => route.scenarioId) });
+    };
+    const setScenario = (scenarioId, emit = true, sync = true) => {
+      clearScenario();
+      if (!scenarioId) {
+        if (emit) wrapper.dispatchEvent(new CustomEvent('cfx-topology-scenario-clear', { bubbles: true, detail: { chartId: attr(wrapper, 'data-chart-id') } }));
+        if (sync) emitSync('scenario', { scenarioId: '' });
+        return;
+      }
+      const route = scenarioRoute(scenarioId);
+      if (!route) {
+        if (emit) wrapper.dispatchEvent(new CustomEvent('cfx-topology-scenario-clear', { bubbles: true, detail: { chartId: attr(wrapper, 'data-chart-id') } }));
+        if (sync) emitSync('scenario', { scenarioId: '' });
+        return;
+      }
+
+      wrapper.setAttribute('data-cfx-active-scenario', scenarioId);
+      if (scenarioControlMode === 'checkboxes') {
+        wrapper.setAttribute('data-cfx-active-scenarios', scenarioId);
+        wrapper.querySelectorAll('[data-cfx-topology-scenario-toggle]').forEach(input => input.checked = attr(input, 'data-cfx-topology-scenario-toggle') === scenarioId);
+      }
+      if (route.color) wrapper.style.setProperty('--cfx-topology-active-scenario-color', route.color);
+      else wrapper.style.removeProperty('--cfx-topology-active-scenario-color');
+      wrapper.querySelectorAll('[data-cfx-topology-scenario]').forEach(button => button.setAttribute('aria-pressed', attr(button, 'data-cfx-topology-scenario') === scenarioId ? 'true' : 'false'));
+      wrapper.querySelectorAll('[data-cfx-role="topology-edge"]').forEach(edge => {
+        const active = route.edgeIds.has(attr(edge, 'data-edge-id'));
+        edge.classList.toggle('cfx-topology-html-scenario-active', active);
+        edge.classList.toggle('cfx-topology-html-scenario-muted', !active);
+        if (active) edge.setAttribute('data-cfx-active-scenario-step-indices', scenarioStepIndices(edge, scenarioId).join(' '));
+      });
+      wrapper.querySelectorAll('[data-cfx-role="topology-node"]').forEach(node => {
+        const active = route.nodeIds.has(attr(node, 'data-node-id'));
+        node.classList.toggle('cfx-topology-html-scenario-active', active);
+        node.classList.toggle('cfx-topology-html-scenario-muted', !active);
+        if (active) node.setAttribute('data-cfx-active-scenario-step-indices', scenarioStepIndices(node, scenarioId).join(' '));
+      });
+      wrapper.querySelectorAll('[data-cfx-role="topology-edge-label"]').forEach(label => {
+        const active = route.edgeIds.has(attr(label, 'data-edge-id'));
+        label.classList.toggle('cfx-topology-html-scenario-active', active);
+        label.classList.toggle('cfx-topology-html-scenario-muted', !active);
+      });
+      wrapper.querySelectorAll('[data-cfx-role="topology-node-status"]').forEach(status => {
+        const active = route.nodeIds.has(attr(status, 'data-node-id'));
+        status.classList.toggle('cfx-topology-html-scenario-active', active);
+        status.classList.toggle('cfx-topology-html-scenario-muted', !active);
+      });
+      wrapper.querySelectorAll('[data-cfx-role="topology-group"]').forEach(group => {
+        const active = route.groupIds.has(attr(group, 'data-group-id'));
+        group.classList.toggle('cfx-topology-html-scenario-active', active);
+        group.classList.toggle('cfx-topology-html-scenario-muted', !active);
+      });
+      renderScenarioPanel(route);
+      clearScenarioStep();
+      updateScenarioStepControls(false);
+      syncScenarioUrl(scenarioId, '');
+      if (emit) wrapper.dispatchEvent(new CustomEvent('cfx-topology-scenario', { bubbles: true, detail: { chartId: attr(wrapper, 'data-chart-id'), scenarioId, label: route.label, description: route.description, color: route.color, stepCount: route.stepCount, metadata: route.metadata, steps: route.steps, nodeIds: Array.from(route.nodeIds), edgeIds: Array.from(route.edgeIds), groupIds: Array.from(route.groupIds) } }));
+      if (sync) emitSync('scenario', { scenarioId });
+    };
+    const svgElement = () => topologySvg();
+    const exportName = () => (attr(wrapper, 'data-chart-id') || 'topology').replace(/[^a-z0-9_.-]+/gi, '-').replace(/^-+|-+$/g, '') || 'topology';
+    const serializeSvg = () => {
+      const svg = svgElement();
+      if (!svg) return null;
+      const clone = svg.cloneNode(true);
+      const defs = clone.querySelector('defs') || clone.insertBefore(document.createElementNS('http://www.w3.org/2000/svg', 'defs'), clone.firstChild);
+      const style = document.createElementNS('http://www.w3.org/2000/svg', 'style');
+      style.setAttribute('data-cfx-export-style', 'force-filters');
+      style.textContent = '.cfx-topology-html-force-hidden{display:none!important}';
+      defs.appendChild(style);
+      return { svg, data: new XMLSerializer().serializeToString(clone) };
+    };
+    const emitExport = format => {
+      wrapper.dispatchEvent(new CustomEvent('cfx-topology-export', { bubbles: true, detail: { chartId: attr(wrapper, 'data-chart-id'), format } }));
+    };
+    const downloadBlob = (blob, extension, format) => {
+      const url = URL.createObjectURL(blob);
+      const link = document.createElement('a');
+      link.href = url;
+      link.download = exportName() + '.' + extension;
+      document.body.appendChild(link);
+      link.click();
+      link.remove();
+      URL.revokeObjectURL(url);
+      emitExport(format);
+    };
+    const svgSize = svg => {
+      const viewBox = svg.viewBox && svg.viewBox.baseVal ? svg.viewBox.baseVal : null;
+      const rect = svg.getBoundingClientRect ? svg.getBoundingClientRect() : { width: 0, height: 0 };
+      const width = Math.max(1, Math.round((viewBox && viewBox.width) || (svg.width && svg.width.baseVal ? svg.width.baseVal.value : 0) || rect.width || 800));
+      const height = Math.max(1, Math.round((viewBox && viewBox.height) || (svg.height && svg.height.baseVal ? svg.height.baseVal.value : 0) || rect.height || 450));
+      return { width, height };
+    };
+    const exportSvg = () => {
+      const serialized = serializeSvg();
+      if (!serialized) return;
+      downloadBlob(new Blob([serialized.data], { type: 'image/svg+xml;charset=utf-8' }), 'svg', 'svg');
+    };
+    const exportPng = () => {
+      const serialized = serializeSvg();
+      if (!serialized) return;
+      const source = new Blob([serialized.data], { type: 'image/svg+xml;charset=utf-8' });
+      const sourceUrl = URL.createObjectURL(source);
+      const image = new Image();
+      image.onload = () => {
+        const size = svgSize(serialized.svg);
+        const scale = Math.max(1, Math.ceil(window.devicePixelRatio || 1));
+        const canvas = document.createElement('canvas');
+        canvas.width = size.width * scale;
+        canvas.height = size.height * scale;
+        const context = canvas.getContext('2d');
+        if (!context) {
+          URL.revokeObjectURL(sourceUrl);
+          return;
+        }
+        context.setTransform(scale, 0, 0, scale, 0, 0);
+        context.clearRect(0, 0, size.width, size.height);
+        context.drawImage(image, 0, 0, size.width, size.height);
+        canvas.toBlob(blob => {
+          URL.revokeObjectURL(sourceUrl);
+          if (blob) downloadBlob(blob, 'png', 'png');
+        }, 'image/png');
+      };
+      image.onerror = () => URL.revokeObjectURL(sourceUrl);
+      image.src = sourceUrl;
+    };
+    const edgeIdentity = edge => ({
+      id: attr(edge, 'data-edge-id'),
+      kind: attr(edge, 'data-edge-kind'),
+      status: attr(edge, 'data-cfx-status'),
+      sourceNodeId: attr(edge, 'data-source-node-id'),
+      targetNodeId: attr(edge, 'data-target-node-id'),
+      sourceGroupId: attr(edge, 'data-source-group-id'),
+      targetGroupId: attr(edge, 'data-target-group-id')
+    });
+    const related = detail => {
+      if (detail.kind === 'node') {
+        const edges = Array.from(wrapper.querySelectorAll('[data-cfx-role="topology-edge"]'))
+          .filter(edge => attr(edge, 'data-source-node-id') === detail.id || attr(edge, 'data-target-node-id') === detail.id);
+        return {
+          nodeIds: unique(edges.flatMap(edge => [attr(edge, 'data-source-node-id'), attr(edge, 'data-target-node-id')]).filter(id => id !== detail.id)),
+          edgeIds: edges.map(edge => attr(edge, 'data-edge-id')),
+          groupIds: unique([detail.groupId].concat(edges.flatMap(edge => [attr(edge, 'data-source-group-id'), attr(edge, 'data-target-group-id')]))),
+          edges: edges.map(edgeIdentity)
+        };
+      }
+      if (detail.kind === 'edge') {
+        return {
+          nodeIds: unique([detail.sourceNodeId, detail.targetNodeId]),
+          edgeIds: [detail.id],
           groupIds: unique([detail.sourceGroupId, detail.targetGroupId]),
           edges: [edgeIdentity(detail.element)]
         };
@@ -335,6 +915,16 @@ public sealed partial class TopologyHtmlRenderer
       if (emit) wrapper.dispatchEvent(new CustomEvent('cfx-topology-select', { bubbles: true, detail }));
       if (sync) emitSync('selection', { selection: { kind: detail.kind, id: detail.id, status: detail.status } });
     };
+    const hydrateSelected = element => {
+      const detail = identity(element);
+      detail.related = related(detail);
+      wrapper.setAttribute('data-cfx-selection-kind', detail.kind);
+      wrapper.setAttribute('data-cfx-selection-id', detail.id);
+      wrapper.setAttribute('data-cfx-selection-status', detail.status);
+      element.setAttribute('aria-selected', 'true');
+      renderSelectionPanel({ ...detail, element });
+      delete detail.element;
+    };
     const forceSearchText = element => [
       attr(element, 'data-node-id'), attr(element, 'data-node-label'), attr(element, 'data-node-kind'), attr(element, 'data-group-id'),
       attr(element, 'data-edge-id'), attr(element, 'data-edge-label'), attr(element, 'data-edge-secondary-label'), attr(element, 'data-edge-tertiary-label'), attr(element, 'data-edge-kind'), attr(element, 'data-source-node-id'), attr(element, 'data-target-node-id'),
@@ -473,7 +1063,7 @@ public sealed partial class TopologyHtmlRenderer
       });
     });
     const initiallySelected = Array.from(wrapper.querySelectorAll(selectables)).find(element => attr(element, 'data-cfx-selected') === 'true');
-    if (initiallySelected) select(initiallySelected, false, false);
+    if (initiallySelected) hydrateSelected(initiallySelected);
     else renderSelectionPanel(null);
     if (viewportControls) {
       applyViewport(viewportState());
@@ -681,6 +1271,3 @@ public sealed partial class TopologyHtmlRenderer
     });
   }
 })();
-</script>
-""";
-}
